First version to calculate the fitness score
I was going to find the collision shapes x layout and add them to a hashmap in order to find the one that has the highest x. The collision that has the highest x value occurs when a car y hits an obstacle, therefore it was perfect to find the car that had the best neural network. However, I know that if the map overlaps with itself my method won't work. So now I think about trying with the distance travelled but I'm not sure how to calculate the time, either with a stopwatch object or the System.nanoTime(). The idea of having checkpoints was suggested by EnYi but I think that it'll be more complicated to implement since there must be multiple points but that it's a great idea. I'll commit what I have for now and think about it later at night.